Refactor assetPath imports for consistency across components
